api/requirements.txt
fastapi
uvicorn
numpy
Pillow
requests
python-multipart